TextClassification/requirement.txt
autopep8==2.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
ffmpeg-python==0.2.0
filelock==3.10.7
fonttools==4.39.3
future==0.18.3
idna==3.4
importlib-metadata==6.1.0
importlib-resources==5.12.0
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
llvmlite==0.39.1
MarkupSafe==2.1.2
matplotlib==3.7.1
more-itertools==9.1.0
mpmath==1.3.0
networkx==3.1
nltk==3.8.1
numba==0.56.4
numpy==1.23.5
openai-whisper @ git+https://github.com/openai/whisper.git@b5851c6c40e753606765ac45b85b298e3ae9e00d
packaging==23.0
Pillow==9.5.0
pycodestyle==2.10.0
pydub==0.25.1
pyparsing==3.0.9
python-dateutil==2.8.2
regex==2023.3.23
requests==2.28.2
Sastrawi==1.0.1
six==1.16.0
sympy==1.11.1
tiktoken==0.3.1
tomli==2.0.1
torch==2.0.0
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
zipp==3.15.0
